Refactor of folder stucture
