+++ mcp_client/registrar.py
