+++ nyxstrike_mcp.py
@@ -12,8 +12,8 @@ Framework: FastMCP integration for tool orchestration
 import sys
 import logging
 from backend.server_core.colored_formatter import ColoredFormatter
-from mcp_client.mcp_core.mcp_entry import run_mcp
-from mcp_client.mcp_core.args import parse_args
+from mcp_client.mcp_entry import run_mcp
+from mcp_client.args import parse_args
 
 # Setup logging
 logging.basicConfig(
